A demo of RISCV in five pipeline stages in sv, and some basic ISA decoder in py.
A.basic riscv design.
In 1 folder: How to decode the ISA instructions to binary sequences that need to be stored in a Instruction Memory (imem).

In 2th foler: the riscv design written in sv,. The top in the top level wrapper that includes imem, dmem and riscv core. The core has controller, datapath and hazard unit. The core has 1 ALU and is capable to do 27 instructions.